+++ src/components/index.ts
@@ -6,4 +6,6 @@ export * from './FullScreenLoader';
 export * from './Icons';
 export * from './LoadingCard';
 export * from './Modal';
+export * from './Pagination';
+export * from './Table';
 export * from './TypeaheadSelect';